+++ packages/fabric/editor-core/__tests__/editor/plugins/table/index.ts
@@ -1,0 +1,14 @@
+import { name } from '../../../../package.json';
+import { tablePlugin } from '../../../../src/editor/plugins';
+
+describe(name, () => {
+  describe('Plugins -> Table', () => {
+    it('should be defined', () => {
+      expect(tablePlugin).toBeDefined();
+    });
+
+    it('should have primaryToolbarComponent', () => {
+      expect(tablePlugin.primaryToolbarComponent).toBeDefined();
+    });
+  });
+});