+++ packages/fabric/editor-core/docs/0-intro.tsx
@@ -2,7 +2,7 @@
 import { md, code } from '@atlaskit/docs';
 
 export default md`
-# Installing @atlaskit/editor
+# Installing @atlaskit/editor-core
 
 \`\`\`
 npm install @atlaskit/editor-core