+++ packages/fabric/editor-core/src/editor/plugins/text-formatting/index.tsx
@@ -64,9 +64,15 @@ const textFormatting = (options: TextFormattingOptions): EditorPlugin => ({
   pmPlugins() {
     return [
       { rank: 800, plugin: () => textFormattingPlugin },
-      { rank: 810, plugin: schema => textFormattingInputRulePlugin(schema) },
+      {
+        rank: 810,
+        plugin: ({ schema }) => textFormattingInputRulePlugin(schema),
+      },
       { rank: 820, plugin: () => clearFormattingPlugin },
-      { rank: 830, plugin: schema => clearFormattingKeymapPlugin(schema) },
+      {
+        rank: 830,
+        plugin: ({ schema }) => clearFormattingKeymapPlugin(schema),
+      },
     ];
   },
 
@@ -89,7 +95,7 @@ const textFormatting = (options: TextFormattingOptions): EditorPlugin => ({
 
     return (
       <ButtonsGroup
-        width={editorWidth! > EditorWidth.BreakPoint6 ? 'large' : 'small'}
+        width={editorWidth! > EditorWidth.BreakPoint10 ? 'large' : 'small'}
       >
         <ToolbarTextFormatting
           disabled={disabled}