+++ packages/fabric/editor-core/src/ui/ChromeExpanded/index.tsx
@@ -15,6 +15,7 @@ import {
 import { BlockTypeState } from '../../plugins/block-type';
 import { CodeBlockState } from '../../plugins/code-block';
 import { stateKey as emojiPluginKey } from '../../plugins/emojis';
+import { stateKey as mediaPluginKey } from '../../plugins/media';
 import { HyperlinkState } from '../../plugins/hyperlink';
 import { ImageUploadState } from '../../plugins/image-upload';
 import { ListsState } from '../../plugins/lists';
@@ -39,6 +40,7 @@ import ToolbarFeedback from '../ToolbarFeedback';
 import ToolbarHelp from '../ToolbarHelp';
 import ToolbarHyperlink from '../ToolbarHyperlink';
 import ToolbarLists from '../ToolbarLists';
+import ToolbarMedia from '../ToolbarMedia';
 import ToolbarTextFormatting from '../ToolbarTextFormatting';
 import ToolbarAdvancedTextFormatting from '../ToolbarAdvancedTextFormatting';
 import ToolbarInsertBlock from '../ToolbarInsertBlock';
@@ -299,6 +301,13 @@ export default class ChromeExpanded extends PureComponent<Props, State> {
                 numFollowingButtons={4}
               />
             )}
+          {pluginStateMedia && (
+            <ToolbarMedia
+              isDisabled={disabled}
+              editorView={editorView}
+              pluginKey={mediaPluginKey}
+            />
+          )}
           {(pluginStateTable ||
             pluginStateMedia ||
             pluginStateBlockType ||