+++ packages/fabric/media-card/examples/card-list-custom-empty-state.tsx
@@ -1,11 +1,6 @@
 import * as React from 'react';
-import {
-  StoryList,
-  createStorybookContext,
-  collectionNames,
-  defaultCollectionName,
-} from '@atlaskit/media-test-helpers';
-import { CardList, CardListEvent } from '../src';
+import { createStorybookContext } from '@atlaskit/media-test-helpers';
+import { CardList } from '../src';
 import { cardsActions, wrongCollection } from '../example-helpers';
 
 const context = createStorybookContext();