+++ packages/fabric/media-card/examples/card-media-type-matrix.tsx
@@ -1,41 +1,20 @@
 import * as React from 'react';
 import {
-  StoryList,
   Matrix,
   createStorybookContext,
-  defaultCollectionName as collectionName,
   videoUrlPreviewId,
   audioUrlPreviewId,
   imageUrlPreviewId,
   docUrlPreviewId,
   unknownUrlPreviewId,
-  genericUrlPreviewId,
-  youTubeUrlPreviewId,
-  spotifyUrlPreviewId,
-  soundcloudUrlPreviewId,
-  publicTrelloBoardUrlPreviewId,
-  privateTrelloBoardUrlPreviewId,
-  errorLinkId,
   videoFileId,
   audioFileId,
   imageFileId,
   docFileId,
   unknownFileId,
-  smallImageFileId,
-  wideImageFileId,
-  largeImageFileId,
-  errorFileId,
 } from '@atlaskit/media-test-helpers';
 
-import {
-  Card,
-  UrlPreviewIdentifier,
-  FileIdentifier,
-  Identifier,
-  CardAppearance,
-  CardEvent,
-  OnSelectChangeFuncResult,
-} from '../src';
+import { Card } from '../src';
 
 const context = createStorybookContext();
 // link cards